# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SMS API",
        "version": "1.0"
    },
    "paths": {
        "/api": {},
        "/auth/token/generate/refresh-token": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh token",
                "description": "RefreshToken",
                "operationId": "refreshToken",
                "requestBody": {
                    "description": "Refresh token",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "test@test.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "type": {
                                            "type": "string",
                                            "example": "refresh"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/UserResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "SmsStatus": {
                "title": "SmsStatus",
                "description": "Sms status",
                "properties": {
                    "WAITING": {
                        "type": "integer",
                        "example": 0
                    },
                    "SENT": {
                        "type": "integer",
                        "example": 1
                    },
                    "FAILED": {
                        "type": "integer",
                        "example": 2
                    }
                },
                "type": "string",
                "enum": [
                    "WAITING",
                    "SENT",
                    "FAILED"
                ]
            },
            "UserResource": {
                "title": "UserResource",
                "description": "User resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "test@test.com"
                    }
                },
                "type": "object"
            },
            "SMS": {
                "title": "SMS",
                "description": "SMS model",
                "properties": {
                    "id": {
                        "description": "SMS ID",
                        "type": "integer",
                        "example": "1"
                    },
                    "user_id": {
                        "description": "User ID",
                        "type": "integer",
                        "example": "1"
                    },
                    "number": {
                        "description": "SMS number",
                        "type": "string",
                        "example": "905555555555"
                    },
                    "message": {
                        "description": "SMS message",
                        "type": "string",
                        "example": "Hello world"
                    },
                    "send_time": {
                        "description": "SMS send time",
                        "type": "string",
                        "example": "2021-01-01 00:00:00"
                    },
                    "status": {
                        "$ref": "#/components/schemas/SmsStatus"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "description": "User name",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "description": "User email",
                        "type": "string",
                        "format": "email",
                        "example": "test@test.com"
                    },
                    "created_at": {
                        "description": "User created at",
                        "type": "string",
                        "example": "2021-01-01 00:00:00"
                    },
                    "updated_at": {
                        "description": "User updated at",
                        "type": "string",
                        "example": "2021-01-01 00:00:00"
                    }
                },
                "type": "object"
            }
        }
    }
}